+++ focad_spatial_location_data.cpp
@@ -7,7 +7,7 @@ FLAMEGPU_AGENT_FUNCTION(focad_spatial_location_data, flamegpu::MessageNone, flam
   FLAMEGPU->message_out.setVariable<float>("fy", FLAMEGPU->getVariable<float>("fy"));
   FLAMEGPU->message_out.setVariable<float>("fz", FLAMEGPU->getVariable<float>("fz"));
   FLAMEGPU->message_out.setVariable<int>("fnode_id", FLAMEGPU->getVariable<int>("fnode_id"));
-  FLAMEGPU->message_out.setVariable<uint8_t>("attached", FLAMEGPU->getVariable<uint8_t>("attached"));
+  FLAMEGPU->message_out.setVariable<int>("attached", FLAMEGPU->getVariable<int>("attached"));
   FLAMEGPU->message_out.setVariable<uint8_t>("active", FLAMEGPU->getVariable<uint8_t>("active"));
